requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-ai==0.0.14
griffe==1.5.1
google-generativeai==0.8.3
python-dotenv==1.0.1
openai>=1.0.0
requests>=2.0.0
